Rewrite the Hyperlog#batch function.
After fixing multiple bugs in this very complex function, I decided to
rewrite it using a literate programming approach. Each step is written
in plain english, and reads from top to bottom. Each step and its
sub-steps are broken up into functions that are, as much as possible,
free of side effects.